Add several error types.
Adds the types JSONMarshalError, JSONUnmarshalError, HTTPRequestError,
HTTPResposeError, and UnmarshalError. These errors represent the
different steps in the GraphQL request where errors can occur.

Change the Errors struct to be exported by the package. Add 'Path'
and 'Type' fields to the Errors struct. The 'path' field is in the
GraphQL specification. The 'type' field is not in the GraphQL
specification. 'type' is returned by the GitHub V4 API.